# swagger.yaml
openapi: 3.0.0
info:
  title: Simple Auth & Dashboard Auth/Data API
  description: Simple Auth & Dashboard is a demonstration web app which features user authentication and a simple dashboard displaying user sessions data. The app is created on the Next.js/Express frameworks and employs Auth0 for user authentication. All API calls are made via the Next.js '/api' routes to Auth0 or to the server's '/api/v1' end point, which require Auth0-issued access token to authenticate access. This documentation includes, besides app API, Auth0 API end points for access token (as necessary for access to app API) and for manual user sign-up with Auth0. It appears Auth0 requires user sign-up/login only via browser authentication flow.
  contact:
    name: Chung Kao
    email: chungckao@gmail.com
  version: "1.0"
servers:
  - url: http://localhost:3000/api
    description: development BaseUrl
  - url: https://simple-auth-and-dashboard.web.app/api
    description: poduction BaseUrl
  - url: http://localhost:3001/api/v1
    description: development ApiUrl
  - url: https://simple-auth-and-dashboard.onrender.com/api/v1
    description: production ApiUrl
  - url: https://dev-lp2pqg88am2bfulk.us.auth0.com
    description: Auth0Url for Auth0 API
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Create User
      operationId: CreateUser
      description: Create a new user
      parameters: []
      requestBody:
        description: auth0_id, profile_name, email, created_at, and logins_count are required and profile_name and email must be unique
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              auth0_id: auth0|123456789101213
              profile_name: johndoe
              email: johndoe@gmail.com
              created_at: 2023-01-23T00:00:01.001Z
              logins_count: 9
        required: true
      responses:
        "201":
          description: Created - returns the new user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
              example:
                msg: User was created successfully
                user:
                  id: 5
                  auth0_id: auth0|123456789101213
                  profile_name: johndoe
                  email: johndoe@gmail.com
                  created_at: 2023-01-23T00:00:01.001Z
                  logins_count: 9
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Users
      summary: Get All Users
      operationId: GetAllUsers
      description: Get all the existing users
      parameters: []
      responses:
        "200":
          description: OK - returns all the users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllUsersResponse"
              example:
                users:
                  - id: 5
                    auth0_id: auth0|123456789101213
                    profile_name: johndoe
                    email: johndoe@gmail.com
                    created_at: 2023-01-23T00:00:01.001Z
                    logins_count: 9
                  - id: 6
                    auth0_id: facebook|123456789111213
                    profile_name: janesmith
                    email: janesmith@gmail.com
                    created_at: 2023-01-24T12:10:24.781Z
                    logins_count: 5
                count: 2
      deprecated: false
      security:
        - bearer: []
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The Auth0 ID for the user
    get:
      tags:
        - Users
      summary: Get Single User
      operationId: GetSingleUser
      description: Get one user by Auth0 ID
      parameters:
        - name: id
          in: path
          description: The Auth0 user ID
          required: true
          style: simple
          schema:
            type: string
            example: auth0|123456789101213
      responses:
        "200":
          description: OK - returns the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSingleUserResponse"
              example:
                user:
                  id: 5
                  auth0_id: auth0|123456789101213
                  profile_name: johndoe
                  email: johndoe@gmail.com
                  created_at: 2023-01-23T00:00:01.001Z
                  logins_count: 9
      deprecated: false
      security:
        - bearer: []
    patch:
      tags:
        - Users
      summary: Update User
      operationId: UpdateUser
      description: Update one user by Auth0 ID
      parameters:
        - name: id
          in: path
          description: The Auth0 user ID
          required: true
          style: simple
          schema:
            type: string
            example: auth0|123456789101213
      requestBody:
        description: Can update either profile_name or logins_count in a single request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              profile_name: John Doe
        required: true
      responses:
        "200":
          description: OK - returns the updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserResponse"
              example:
                msg: User was updated successfully
                user:
                  id: 5
                  auth0_id: auth0|123456789101213
                  profile_name: John Doe
                  email: johndoe@gmail.com
                  created_at: 2023-01-23T00:00:01.001Z
                  logins_count: 8
      deprecated: false
      security:
        - bearer: []
    delete:
      tags:
        - Users
      summary: Delete User
      operationId: DeleteUser
      description: For admin only - delete one user by Auth0 ID
      parameters:
        - name: id
          in: path
          description: Auth0 user ID
          required: true
          style: simple
          schema:
            type: string
            example: auth0|123456789101213
      responses:
        "200":
          description: OK - returns the deleted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteUserResponse"
              example:
                msg: User was removed successfully
                user:
                  id: 5
                  auth0_id: auth0|123456789101213
                  profile_name: John Doe
                  email: johndoe@gmail.com
                  created_at: 2023-01-23T00:00:01.001Z
                  logins_count: 8
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    CreateUserRequest:
      title: CreateUserRequest
      required:
        - auth0_id
        - profile_name
        - email
        - created_at
        - logins_count
      unique:
        - profile_name
      type: object
      properties:
        auth0_id:
          type: string
        profile_name:
          type: string
        email:
          type: string
        created_at:
          type: string
        logins_count:
          type: number
      example:
        auth0_id: auth0|123456789101213
        profile_name: johndoe
        email: johndoe@gmail.com
        created_at: 2023-01-23T00:00:01.001Z
        logins_count: 9
    CreateUserResponse:
      title: CreateUserResponse
      type: object
      properties:
        msg:
          type: string
        user:
          type: object
          properties:
            id:
              type: number
            auth0_id:
              type: string
            profile_name:
              type: string
            email:
              type: string
            created_at:
              type: string
            logins_count:
              type: number
      example:
        msg: User was created successfully
        user:
          id: 5
          auth0_id: auth0|123456789101213
          profile_name: johndoe
          email: johndoe@gmail.com
          created_at: 2023-01-23T00:00:01.001Z
          logins_count: 9
    GetAllUsersResponse:
      title: GetAllUsersResponse
      type: object
      properties:
        users:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              auth0_id:
                type: string
              profile_name:
                type: string
              email:
                type: string
              created_at:
                type: string
              logins_count:
                type: number
        count:
          type: number
      example:
        users:
          - id: 5
            auth0_id: auth0|123456789101213
            profile_name: johndoe
            email: johndoe@gmail.com
            created_at: 2023-01-23T00:00:01.001Z
            logins_count: 9
          - id: 6
            auth0_id: facebook|123456789111213
            profile_name: janesmith
            email: janesmith@gmail.com
            created_at: 2023-01-24T12:10:24.781Z
            logins_count: 5
        count: 15
    GetSingleUserResponse:
      title: GetSingleUserResponse
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: number
            auth0_id:
              type: string
            profile_name:
              type: string
            email:
              type: string
            created_at:
              type: string
            logins_count:
              type: number
      example:
        user:
          id: 5
          auth0_id: auth0|123456789101213
          profile_name: johndoe
          email: johndoe@gmail.com
          created_at: 2023-01-23T00:00:01.001Z
          logins_count: 9
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
        - profile_name
        - logins_count
      type: object
      properties:
        profile_name:
          type: string
        logins_count:
          type: number
      example:
        profile_name: John Doe
        logins_count: 8
    UpdateUserResponse:
      title: UpdateUserResponse
      type: object
      properties:
        msg:
          type: string
        user:
          type: object
          properties:
            id:
              type: number
            auth0_id:
              type: string
            profile_name:
              type: string
            email:
              type: string
            created_at:
              type: string
            logins_count:
              type: number
      example:
        msg: User was updated successfully
        user:
          id: 5
          auth0_id: auth0|123456789101213
          profile_name: John Doe
          email: johndoe@gmail.com
          created_at: 2023-01-23T00:00:01.001Z
          logins_count: 8
    DeleteUserResponse:
      title: DeleteUserResponse
      type: object
      properties:
        msg:
          type: string
        user:
          type: object
          properties:
            id:
              type: number
            auth0_id:
              type: string
            profile_name:
              type: string
            email:
              type: string
            created_at:
              type: string
            logins_count:
              type: number
      example:
        msg: User was removed successly
        user:
          id: 5
          auth0_id: auth0|123456789101213
          profile_name: John Doe
          email: johndoe@gmail.com
          created_at: 2023-01-23T00:00:01.001Z
          Logins_count: 8
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...